Move build code to build_menu
